private route added

+++ src/Pages/Jobs/Jobs.jsx
@@ -1,14 +1,18 @@
-import axios from "axios";
 import { useEffect, useState } from "react";
 import Job from "../../Components/Job/Job";
 import "./Jobs.css";
 import ApplyJobSide from "../../Components/ApplyJobSide/ApplyJobSide";
 import AppliedJobsSide from "../../Components/AppliedJobsSide/AppliedJobsSide";
 import Loading from "../../Components/Loading/Loading";
+import { onAuthStateChanged } from "firebase/auth";
+import auth from "../../Config/Firebase";
+import { NavLink, useNavigate } from "react-router-dom";
+import axios from "axios";
 
 const Jobs = () => {
   const [jobs, setJobs] = useState([]);
   const URL = "http://localhost:9000/jobs";
+  const [isUser, setIsUser] = useState(false);
 
   useEffect(() => {
     axios.get(URL).then((response) => {
@@ -17,17 +21,26 @@ const Jobs = () => {
   }, [URL]);
 
   const handleDelete = (id) => {
-    axios.delete(`${URL}/${id}`).then(() => {
-      confirm("Do you want to delete the post!");
-      setJobs((prevJobs) => prevJobs.filter((job) => job.id !== id));
-    });
+    const shouldDelete = window.confirm("Do you want to delete the post!");
+
+    if (shouldDelete) {
+      axios.delete(`${URL}/${id}`).then(() => {
+        setJobs((prevJobs) => prevJobs.filter((job) => job.id !== id));
+      });
+    }
   };
+
+  onAuthStateChanged(auth, (user) => {
+    if (user) {
+      setIsUser(true);
+    }
+  });
 
   if (!jobs) {
     return <Loading />;
   }
 
-  if (!jobs || jobs.length === 0) {
+  if (jobs.length === 0) {
     return (
       <>
         <h1>No jobs added!</h1>
@@ -44,7 +57,19 @@ const Jobs = () => {
             <Job key={job.id} job={job} handleDelete={handleDelete} />
           ))}
         </div>
-        <AppliedJobsSide />
+        {isUser ? (
+          <AppliedJobsSide />
+        ) : (
+          <div className="no-applied">
+            <p>Please log in to see applied jobs!</p>
+            <NavLink to="/login">
+              {" "}
+              <button className="primary-btn " id="no-applied">
+                Load Applied Jobs
+              </button>{" "}
+            </NavLink>
+          </div>
+        )}
       </div>
     </>
   );
